+++ bin/firebase_verify_token_dart.dart
@@ -1,17 +1,10 @@
 import 'package:firebase_verify_token_dart/firebase_verify_token_dart.dart';
 
 void main() async {
-  FirebaseVerifyToken.projectIds = ['test', 'test-1'];
+  FirebaseVerifyToken.projectIds = ['cbes-c64d6', 'test-1'];
 
-  final result = await FirebaseVerifyToken.verify(
-    'token',
-    onVerifySuccessful: ({required bool status, String? projectId}) {
-      print('STATUS: $status');
-      print('PROJECT ID: $projectId');
-    },
-  );
   await FirebaseVerifyToken.verify(
-    'my-token-string',
+    'token',
     onVerifySuccessful: ({required bool status, String? projectId}) {
       if (status) {
         print('Token verified for project: $projectId');
@@ -20,9 +13,4 @@ void main() async {
       }
     },
   );
-  if (result) {
-    print('TOKEN IS VERIFIED');
-  } else {
-    print('TOKEN IS NOT VERIFIED');
-  }
 }